+++ dio_cache_interceptor/lib/dio_cache_interceptor.dart
@@ -3,3 +3,6 @@ export 'package:http_cache_core/src/store/store.dart';
 
 export 'src/dio_cache_interceptor.dart';
 export 'src/extension/cache_option_extension.dart';
+export 'src/extension/cache_response_extension.dart';
+export 'src/extension/request_extension.dart';
+export 'src/extension/response_extension.dart';