+++ dio_cache_interceptor/lib/src/extension/request_extension.dart
@@ -8,7 +8,7 @@ extension RequestExtension on RequestOptions {
   List<String>? headerValuesAsList(String headerKey) {
     final value = headers[headerKey];
 
-    if (value is List<String>) return value;
+    if (value is Iterable<String>) return value.toList();
     if (value is String) return value.split(',').map((h) => h.trim()).toList();
 
     return value;
@@ -16,5 +16,20 @@ extension RequestExtension on RequestOptions {
 
   CacheOptions? getCacheOptions() {
     return extra[extraKey];
+  }
+
+  /// Get headers flatten to String
+  Map<String, String> getFlattenHeaders() {
+    final h = <String, String>{};
+
+    for (var header in headers.entries) {
+      if (header.value is Iterable<String>) {
+        h[header.key] = header.value.join(',');
+      } else if (header.value is String) {
+        h[header.key] = header.value;
+      }
+    }
+
+    return h;
   }
 }