+++ dio_cache_interceptor/test/cache_interceptor_maxstale_test.dart
@@ -1,6 +1,5 @@
 import 'package:dio/dio.dart';
 import 'package:dio_cache_interceptor/dio_cache_interceptor.dart';
-import 'package:dio_cache_interceptor/src/extension/cache_response_extension.dart';
 import 'package:test/test.dart';
 
 import 'mock_httpclient_adapter.dart';