+++ dio_cache_interceptor/test/cache_interceptor_test.dart
@@ -2,7 +2,6 @@ import 'dart:typed_data';
 
 import 'package:dio/dio.dart';
 import 'package:dio_cache_interceptor/dio_cache_interceptor.dart';
-import 'package:dio_cache_interceptor/src/extension/cache_response_extension.dart';
 import 'package:test/test.dart';
 
 import 'mock_httpclient_adapter.dart';